+++ src/lib.rs
@@ -1,13 +1,6 @@
 // Allow proc-macros to use `ailoy::` prefix when invoked from within this crate.
 extern crate self as ailoy;
 
-pub(crate) mod agent;
-pub(crate) mod datatype;
-pub(crate) mod lang_model;
-pub(crate) mod message;
-pub(crate) mod tool;
-
-pub use agent::*;
-pub use lang_model::*;
-pub use message::*;
-pub use tool::*;
+pub mod agent;
+pub mod datatype;
+pub mod message;
