--- src/tool/knowledge.rs
@@ -1,7 +1,0 @@
-use std::sync::Arc;
-
-use crate::{ToolDesc, ToolFunc};
-
-pub fn create_knowledge_search_tool<'a>() -> (ToolDesc, Arc<ToolFunc<'a>>) {
-    todo!()
-}
